Include LLM explanation in the assessment results
